Relocate the Junit Test for EventTest
Necessary to comply with the file organisation
